feat: Add global graph icon to Graph component
The code changes in this commit add a global graph icon to the Graph component in the `Graph.tsx` file. The icon is added as a child element of the `graph-outer` div with the id `global-graph-icon`. This change enhances the visual representation of the graph component.

Note: This commit message follows the established convention of using a prefix (feat) to indicate a new feature or enhancement.